refactor: dropping the general "Asset" path in favor of "styles/util/ImageAsset" and "styles/util/Slice9Asset"
